Ajout sécurité de suppression

See merge request nozama2023/nozamafront!46

+++ src/app/categories/pages/page-list-categories/page-list-categories.component.ts
@@ -29,6 +29,7 @@ export class PageListCategoriesComponent implements OnInit {
   }
 
   delete(category: Categories) {
+    if(confirm("Voulez vous supprimer la catégorie ?"))
     this.categoriesService.deleteCategoryById(category.id)
       .subscribe(() => {
         this.getData();
